// src/models/RemoteEventDetail.ts
export type RemoteEventDetail = {
    boardIdx: string;
    boardSubIdx: number;
    notice: string;
    memberType: string;
    memberIdx: string;
    memberEmail: string;
    memberName: string;
    hit: number;
    useTag: string;
    createdAt: string;
    boardSubject: string;
    boardContents: string;
    operatingType: string;
    reply: string[];
    imageUrl: string;
};